feat(#1417): move FakeMaven out of package and remove the package itself
