+++ eo-maven-plugin/src/test/java/org/eolang/maven/ParseMojoTest.java
@@ -32,7 +32,6 @@ import org.cactoos.text.TextOf;
 import org.cactoos.text.UncheckedText;
 import org.eolang.maven.hash.ChNarrow;
 import org.eolang.maven.hash.ChRemote;
-import org.eolang.maven.testapi.FakeMaven;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Assertions;
